finished page generation functionality and can serve page on localhost

+++ src/main.py
@@ -1,23 +1,26 @@
 from copystatic import copy_directory_contents_recursive
 from generatepage import extract_title, generate_path
 import shutil
-import sys
+import os
 
 def copy_static():
     #deletes the destination directory
     print("Deleting destination directory...")
     dest_dir = "./public"
     src_dir = "./static"
-    shutil.rmtree(dest_dir)
+    if os.path.exists(dest_dir):
+        shutil.rmtree(dest_dir)
     print("Copying static files to destination directory...")
     copy_directory_contents_recursive(src_dir, dest_dir)
 
-
 def main():
+
     copy_static()
+
     from_path = "./content/index.md"
     template_path = "./template.html"
-    dest_path = "./placeholder.txt"
+    dest_path = "./public/index.html"
+
     generate_path(from_path, template_path, dest_path)
